trying int tensor as input

+++ pepper_variant/modules/python/models/dataloader.py
@@ -140,7 +140,7 @@ class SequenceDatasetFake(Dataset):
         base_label = [item[6] for item in batch]
         type_label = [item[7] for item in batch]
 
-        image = image.type(torch.FloatTensor)
+        image = image.type(torch.IntTensor)
 
         return [contig, position, depth, candidate, candidate_frequency, image, base_label, type_label]
